Rename exception and plugin modules to plural names
