+++ src/snowcli/api/cli_global_context.py
@@ -2,7 +2,7 @@ import re
 from pathlib import Path
 from typing import Dict, Optional
 
-from snowcli.api.exception import InvalidSchemaError
+from snowcli.api.exceptions import InvalidSchemaError
 from snowcli.api.output.formats import OutputFormat
 from snowcli.app.snow_connector import connect_to_snowflake
 from snowflake.connector import SnowflakeConnection
