+++ src/snowcli/api/commands/flags.py
@@ -210,7 +210,7 @@ def execution_identifier_argument(sf_object: str, example: str) -> typer.Argumen
 
 
 def project_definition_option(project_name: str):
-    from snowcli.api.exception import NoProjectDefinitionError
+    from snowcli.api.exceptions import NoProjectDefinitionError
     from snowcli.cli.project.definition_manager import DefinitionManager
 
     def _callback(project_path: Optional[str]):
