+++ src/snowcli/api/exceptions.py
