+++ src/snowcli/app/commands_registration/commands_registration_with_callbacks.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass
 from typing import Callable, List
 
-from snowcli.api.plugin.plugin_config import PluginConfigProvider
+from snowcli.api.plugins.plugin_config import PluginConfigProvider
 from snowcli.app.commands_registration.command_plugins_loader import (
     load_builtin_and_external_command_plugins,
     load_only_builtin_command_plugins,
