+++ src/snowcli/cli/nativeapp/manager.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from textwrap import dedent
 from typing import Dict, List, Optional
 
-from snowcli.api.exception import SnowflakeSQLExecutionError
+from snowcli.api.exceptions import SnowflakeSQLExecutionError
 from snowcli.api.sql_execution import SqlExecutionMixin
 from snowcli.cli.connection.util import make_snowsight_url
 from snowcli.cli.nativeapp.artifacts import (
