+++ src/snowcli/cli/nativeapp/version/version_processor.py
@@ -8,7 +8,7 @@ from click import ClickException
 from git import Repo
 from git.exc import InvalidGitRepositoryError
 from rich import print
-from snowcli.api.exception import SnowflakeSQLExecutionError
+from snowcli.api.exceptions import SnowflakeSQLExecutionError
 from snowcli.cli.nativeapp.artifacts import find_version_info_in_manifest_file
 from snowcli.cli.nativeapp.constants import VERSION_COL
 from snowcli.cli.nativeapp.exceptions import ApplicationPackageDoesNotExistError
